+++ releases.md
@@ -1,5 +1,9 @@
 # Releases
 
+## Unreleased
+
+  - Add support for multi-hop proxying.
+
 ## v0.2.0
 
   - Fix handling of temporary objects.